Temp workaround for build failures
Current implementation of client side encryption requires the encryption keys to be present at build time rather than runtime. This makes distributing and deploying the Docker container impossible. This is a workaround for the failing builds to allow for merging the Postgres/Docker change and then later on reworking client side encryption.